requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
altair==5.3.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
boto3==1.34.132
botocore==1.34.132
cachelib==0.9.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-HTTPAuth==4.8.0
flask-profiler==1.8.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
numpy==2.0.0
openai==1.35.13
packaging==24.1
pandas==2.2.2
pdfminer==20191125
pdfminer.six==20231228
pillow==10.3.0
plotly==5.22.0
prompt_toolkit==3.0.47
protobuf==5.27.1
pyarrow==16.1.0
pycodestyle==2.12.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.2
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
streamlit==1.36.0
tenacity==8.4.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
vine==5.1.0
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.13.1
zipp==3.19.2
